feat: fillter for todo

+++ packages/backend/src/controllers/todo/todo.controller.ts
@@ -1,4 +1,5 @@
 import { TodoService } from '@/services';
+import { STATUSES_KEYS } from '@/types/todos.types';
 import { User } from '@prisma/client';
 import { Response, Request } from 'express';
 
@@ -7,7 +8,15 @@ export class TodoController {
 
 	async getAllTodo(req: Request, res: Response): Promise<void> {
 		const ownerId = (req.user as User)?.id;
-		const todos = await this.todoService.findAll(ownerId);
+		const searchQuery = req.query.search as string;
+		const status = req.query.status as STATUSES_KEYS;
+
+		const todos = await this.todoService.findAll(
+			ownerId,
+			searchQuery,
+			status,
+		);
+
 		res.status(200).json(todos);
 	}
 
@@ -32,6 +41,14 @@ export class TodoController {
 
 		const newTodo = req.body;
 
+		const todo = await this.todoService.findById(id);
+		if (todo?.ownerId !== ownerId) {
+			res.status(403).json({
+				message: 'You do not have permission to edit this todo',
+			});
+			return;
+		}
+
 		const updatedTodo = await this.todoService.updateTodo(
 			ownerId,
 			id,
@@ -44,6 +61,22 @@ export class TodoController {
 	async deleteTodo(req: Request, res: Response): Promise<void> {
 		const { id } = req.params;
 		const ownerId = (req.user as User)?.id;
+		const todo = await this.todoService.findById(id);
+
+		if (!todo) {
+			res.status(404).json({
+				message: 'Todo not found',
+			});
+			return;
+		}
+
+		if (todo.ownerId !== ownerId) {
+			res.status(403).json({
+				message: 'You do not have permission to delete this todo',
+			});
+			return;
+		}
+
 		await this.todoService.deleteTodo(ownerId, id);
 		res.status(204).end();
 	}
